backend/requirements.txt
fastapi
uvicorn

# for database
sqlalchemy
psycopg2

# for env var
python-dotenv

# for email validation
pydantic[email]

# for hashing
passlib[bcrypt]

# for testing
pytest
requests
pytest-cov

# for jwt tokens
python-jose[cryptography]

# for form data
python-multipart